+++ src/main/java/net/finmath/smartcontract/valuation/oracle/ValuationOracle.java
@@ -7,6 +7,7 @@
 package net.finmath.smartcontract.valuation.oracle;
 
 import javax.money.MonetaryAmount;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Map;
 
@@ -26,9 +27,9 @@ public interface ValuationOracle {
 	 * @param marketDataTime The market data time.
 	 * @return The value.
 	 */
-	Double getValue(LocalDateTime evaluationTime, LocalDateTime marketDataTime);
+	BigDecimal getValue(LocalDateTime evaluationTime, LocalDateTime marketDataTime);
 
-	Map<String,Double> getValues(LocalDateTime evaluationTime, LocalDateTime marketDataTime);
+	Map<String,BigDecimal> getValues(LocalDateTime evaluationTime, LocalDateTime marketDataTime);
 
 	/**
 	 * Provides the value of the Oracle at a given evaluation time.
