+++ src/main/java/net/finmath/smartcontract/valuation/oracle/ValuationOracleSamplePath.java
@@ -11,6 +11,7 @@ import org.javamoney.moneta.Money;
 import javax.money.CurrencyUnit;
 import javax.money.Monetary;
 import javax.money.MonetaryAmount;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Map;
 
@@ -39,12 +40,12 @@ public class ValuationOracleSamplePath implements ValuationOracle {
 	}
 
 	@Override
-	public Double getValue(final LocalDateTime evaluationTime, final LocalDateTime marketDataTime) {
-		return stochasticValuationOracle.getValue(evaluationTime, marketDataTime).get(path);
+	public BigDecimal getValue(final LocalDateTime evaluationTime, final LocalDateTime marketDataTime) {
+		return BigDecimal.valueOf(stochasticValuationOracle.getValue(evaluationTime, marketDataTime).get(path));
 	}
 
 	@Override
-	public Map<String, Double> getValues(LocalDateTime evaluationTime, LocalDateTime marketDataTime) {
+	public Map<String, BigDecimal> getValues(LocalDateTime evaluationTime, LocalDateTime marketDataTime) {
 		return Map.of("value", getValue(evaluationTime, marketDataTime));
 	}
